Update kdf_nist.h
Fix an absurd bug on the version 0.2.1

+++ kdf_nist_sp800-108/kdf_nist.h
@@ -3,7 +3,7 @@
  *
  *      Author  : Truman
  *      Contact : truman.t.kim@gmail.com
- *      Version : 0.2.0
+ *      Version : 0.2.1
  */
 
 #ifndef KDF_NIST_H_
@@ -124,18 +124,18 @@
 
 #define DIAGNOSTICS
 
-int KDF_CTR_HMAC(uint8_t *Ko, size_t L,
+int KDF_CTR_HMAC(uint8_t *Ko, uint32_t L,
                         uint8_t *Ki, size_t Ki_len,
                         uint8_t *Label, size_t Label_len,
                         uint8_t *Context, size_t Context_len,
                         uint32_t cfg);
 
-int KDF_CTR_HMAC_SHA256(uint8_t *Ko, size_t L,
+int KDF_CTR_HMAC_SHA256(uint8_t *Ko, uint32_t L,
                         uint8_t *Ki, size_t Ki_len,
                         uint8_t *Label, size_t Label_len,
                         uint8_t *Context, size_t Context_len);
 
-int KDF_CTR_HMAC_SHA512(uint8_t *Ko, size_t L,
+int KDF_CTR_HMAC_SHA512(uint8_t *Ko, uint32_t L,
                         uint8_t *Ki, size_t Ki_len,
                         uint8_t *Label, size_t Label_len,
                         uint8_t *Context, size_t Context_len);
